+++ lib/bitcoinex_explorer_web/controllers/channels_redirect_controller.ex
@@ -1,0 +1,7 @@
+defmodule BitcoinexExplorerWeb.ChannelsRedirectController do
+  use BitcoinexExplorerWeb, :controller
+
+  def redirect_to_home(conn, _params) do
+    redirect(conn, to: ~p"/")
+  end
+end